requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
prometheus-fastapi-instrumentator
python-json-logger
slowapi
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
langchain
langchain-core
langchain-community
langchain-openai
langchain-anthropic
langchain-groq
langchain-google-genai
langchain-mistralai
langchain-text-splitters
redis
rq
chromadb
sentence-transformers
pypdf
pymupdf